Renamed and re-organized Literal -> Object
Rather than returning Node-objects from our builtins we now
return the more-specific "Object"-objects.  This covers our
primitive types perfectly:

* String
* Number
* Boolean
* Funcall

This avoids potential errors more easily, and is logically cleaner.

The objec-implementations have been moved to `ast/object.go` and
all builtin-code to the `ast/builtin.go` file too.
