+++ HACKING.md
@@ -16,7 +16,11 @@ Our implementation is pretty simple and all revolves around a set of rules.
 
 * The other half of our code is involved with executing the rules.
   * The main driver is the [executor](executor/) package, which runs rules.
-  * Conditional execution is managed via the built-in functions in [builtin.go](ast/builtin.go) file.
+  * Conditional execution is managed via the built-in functions located in the [ast/builtin.go](ast/builtin.go) file.
+
+* We use a bunch of objects, stored beneath `ast/` which implement simple primitives
+  * Strings, Numbers, Booleans, and Functions are implemented in [ast/object.go](ast/object.go)
+  * These all have an evaluation-method which allow them to self-execute and return strings.
 
 In addition to the above we also have a [config](config/) object which is passed around to allow us to centralize global state, and we have a set of [file](file/) helpers which contain some central code.
