+++ ast/builtin.go
@@ -1,3 +1,5 @@
+// builtin.go - Contains our built-in primitives.
+
 package ast
 
 import (
@@ -15,9 +17,56 @@ import (
 	"github.com/skx/marionette/file"
 )
 
+// BuiltIn is the signature of a built-in function
+type BuiltIn func(env *environment.Environment, args []string) (Object, error)
+
+// FUNCTIONS contains our list of built-in functions, as a map.
+//
+// The key is the name of the function, and the value is the pointer to the
+// function which is used to implement it.
+var FUNCTIONS map[string]BuiltIn
+
+// FALSE is a global false-value, which simplifies our function returns
+var FALSE = &Boolean{Value: false}
+
+// TRUE is a global true-value, which simplifies our function returns.
+var TRUE = &Boolean{Value: true}
+
+// init is called on startup, and creates the FUNCTIONS map which will
+// hold our built-in functions.
+func init() {
+
+	// create the map to hold function-references
+	FUNCTIONS = make(map[string]BuiltIn)
+
+	// Populate it.
+	FUNCTIONS["contains"] = fnContains
+	FUNCTIONS["empty"] = fnEmpty
+	FUNCTIONS["equal"] = fnEqual
+	FUNCTIONS["equals"] = fnEqual // duplicate
+	FUNCTIONS["exists"] = fnExists
+	FUNCTIONS["failure"] = fnFailure
+	FUNCTIONS["len"] = fnLen
+	FUNCTIONS["lower"] = fnLower
+	FUNCTIONS["md5sum"] = fnMD5Sum
+	FUNCTIONS["md5"] = fnMD5Sum // duplicate
+	FUNCTIONS["nonempty"] = fnNonEmpty
+	FUNCTIONS["on_path"] = fnOnPath
+	FUNCTIONS["set"] = fnNonEmpty // duplicate
+	FUNCTIONS["sha1sum"] = fnSha1Sum
+	FUNCTIONS["sha1"] = fnSha1Sum // duplicate
+	FUNCTIONS["success"] = fnSuccess
+	FUNCTIONS["unset"] = fnEmpty // duplicate
+	FUNCTIONS["upper"] = fnUpper
+}
+
+//
+// Now our built-in methods follow
+//
+
 // fnContains returns true/false depending upon whether the first string
 // contains the second one.
-func fnContains(env *environment.Environment, args []string) (Node, error) {
+func fnContains(env *environment.Environment, args []string) (Object, error) {
 
 	// Two arguments are required.
 	if len(args) != 2 {
@@ -25,28 +74,28 @@ func fnContains(env *environment.Environment, args []string) (Node, error) {
 	}
 
 	if strings.Contains(args[0], args[1]) {
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 	}
 
-	return &Boolean{Value: false}, nil
+	return FALSE, nil
 }
 
 // fnEmpty returns true if the given string is unset/empty
-func fnEmpty(env *environment.Environment, args []string) (Node, error) {
+func fnEmpty(env *environment.Environment, args []string) (Object, error) {
 	if len(args) != 1 {
 		return nil, fmt.Errorf("wrong number of args for 'empty': %d != 1", len(args))
 	}
 
 	if len(args[0]) == 0 {
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 	}
 
-	return &Boolean{Value: false}, nil
+	return FALSE, nil
 }
 
 // fnEqual returns true/false depending upon whether the two arguments
 // are equal.
-func fnEqual(env *environment.Environment, args []string) (Node, error) {
+func fnEqual(env *environment.Environment, args []string) (Object, error) {
 
 	// Two arguments are required.
 	if len(args) != 2 {
@@ -55,28 +104,28 @@ func fnEqual(env *environment.Environment, args []string) (Node, error) {
 
 	// If the values differ then not-equal
 	if args[0] != args[1] {
-		return &Boolean{Value: false}, nil
+		return FALSE, nil
 	}
 
 	// Same values?  Then equal
-	return &Boolean{Value: true}, nil
+	return TRUE, nil
 }
 
 // fnExists takes a single argument, and returns true if the specified file exists.
-func fnExists(env *environment.Environment, args []string) (Node, error) {
+func fnExists(env *environment.Environment, args []string) (Object, error) {
 	if len(args) != 1 {
 		return nil, fmt.Errorf("wrong number of args for 'exists': %d != 1", len(args))
 	}
 
 	if file.Exists(args[0]) {
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 	}
 
-	return &Boolean{Value: false}, nil
+	return FALSE, nil
 }
 
 // fnFailure returns true if executing the given command fails.
-func fnFailure(env *environment.Environment, args []string) (Node, error) {
+func fnFailure(env *environment.Environment, args []string) (Object, error) {
 
 	if len(args) != 1 {
 		return nil, fmt.Errorf("wrong number of args for 'failure': %d != 1", len(args))
@@ -99,25 +148,25 @@ func fnFailure(env *environment.Environment, args []string) (Node, error) {
 	if errors.As(err, &ee) {
 		// ran, but non-zero exit code
 		// log.Println("exit code error:", ee.ExitCode())
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 
 	} else if errors.As(err, &pe) {
 		// "no such file ...", "permission denied" etc.
 		// log.Printf("os.PathError: %v", pe)
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 
 	} else if err != nil {
 		// something really bad happened!
 		//log.Printf("general error: %v", err)
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 	} else {
 		// No failure
-		return &Boolean{Value: false}, nil
+		return FALSE, nil
 	}
 }
 
 // fnLen returns the length of the given node.
-func fnLen(env *environment.Environment, args []string) (Node, error) {
+func fnLen(env *environment.Environment, args []string) (Object, error) {
 
 	// Only one argument is supported.
 	if len(args) != 1 {
@@ -128,7 +177,7 @@ func fnLen(env *environment.Environment, args []string) (Node, error) {
 }
 
 // fnLower converts the given node to lower-case.
-func fnLower(env *environment.Environment, args []string) (Node, error) {
+func fnLower(env *environment.Environment, args []string) (Object, error) {
 
 	// Only one argument is supported.
 	if len(args) != 1 {
@@ -139,7 +188,7 @@ func fnLower(env *environment.Environment, args []string) (Node, error) {
 }
 
 // fnMD5Sum returns the MD5 digest of the given input
-func fnMD5Sum(env *environment.Environment, args []string) (Node, error) {
+func fnMD5Sum(env *environment.Environment, args []string) (Object, error) {
 
 	if len(args) != 1 {
 		return nil, fmt.Errorf("wrong number of args for 'md5sum': %d != 1", len(args))
@@ -152,20 +201,20 @@ func fnMD5Sum(env *environment.Environment, args []string) (Node, error) {
 }
 
 // fnNonEmpty returns true if the given string is not unset/empty
-func fnNonEmpty(env *environment.Environment, args []string) (Node, error) {
+func fnNonEmpty(env *environment.Environment, args []string) (Object, error) {
 	if len(args) != 1 {
 		return nil, fmt.Errorf("wrong number of args for 'nonempty': %d != 1", len(args))
 	}
 
 	if len(args[0]) > 0 {
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 	}
 
-	return &Boolean{Value: false}, nil
+	return FALSE, nil
 }
 
 // fnOnPath returns true if the given binary can be found on the users' PATH
-func fnOnPath(env *environment.Environment, args []string) (Node, error) {
+func fnOnPath(env *environment.Environment, args []string) (Object, error) {
 
 	// only one argument is allowed
 	if len(args) != 1 {
@@ -181,22 +230,22 @@ func fnOnPath(env *environment.Environment, args []string) (Node, error) {
 	// got an error? return it
 	if err != nil {
 		log.Printf("[DEBUG] error running lookup %s", err)
-		return &Boolean{Value: false}, nil
+		return FALSE, nil
 	}
 
 	log.Printf("[DEBUG] lookup resulted in '%s'", path)
 
 	// found a path?  Then success
 	if path != "" {
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 	}
 
 	// Not found
-	return &Boolean{Value: false}, nil
+	return FALSE, nil
 }
 
 // fnSha1Sum returns the SHA1 digest of the given input
-func fnSha1Sum(env *environment.Environment, args []string) (Node, error) {
+func fnSha1Sum(env *environment.Environment, args []string) (Object, error) {
 
 	if len(args) != 1 {
 		return nil, fmt.Errorf("wrong number of args for 'sha1sum': %d != 1", len(args))
@@ -209,7 +258,7 @@ func fnSha1Sum(env *environment.Environment, args []string) (Node, error) {
 }
 
 // fnSuccess returns true if executing the given command succeeds.
-func fnSuccess(env *environment.Environment, args []string) (Node, error) {
+func fnSuccess(env *environment.Environment, args []string) (Object, error) {
 
 	if len(args) != 1 {
 		return nil, fmt.Errorf("wrong number of args for 'success': %d != 1", len(args))
@@ -237,20 +286,20 @@ func fnSuccess(env *environment.Environment, args []string) (Node, error) {
 	} else if errors.As(err, &pe) {
 		// "no such file ...", "permission denied" etc.
 		// log.Printf("os.PathError: %v", pe)
-		return &Boolean{Value: false}, nil
+		return FALSE, nil
 
 	} else if err != nil {
 		// something really bad happened!
 		//log.Printf("general error: %v", err)
-		return &Boolean{Value: false}, nil
+		return FALSE, nil
 	} else {
 		// no error
-		return &Boolean{Value: true}, nil
+		return TRUE, nil
 	}
 }
 
 // fnUpper converts the given node to upper-case.
-func fnUpper(env *environment.Environment, args []string) (Node, error) {
+func fnUpper(env *environment.Environment, args []string) (Object, error) {
 
 	// Only one argument is supported.
 	if len(args) != 1 {
@@ -258,25 +307,4 @@ func fnUpper(env *environment.Environment, args []string) (Node, error) {
 	}
 
 	return &String{Value: strings.ToUpper(args[0])}, nil
-}
-
-func init() {
-	FUNCTIONS["contains"] = fnContains
-	FUNCTIONS["empty"] = fnEmpty
-	FUNCTIONS["equal"] = fnEqual
-	FUNCTIONS["equals"] = fnEqual
-	FUNCTIONS["exists"] = fnExists
-	FUNCTIONS["failure"] = fnFailure
-	FUNCTIONS["len"] = fnLen
-	FUNCTIONS["lower"] = fnLower
-	FUNCTIONS["md5sum"] = fnMD5Sum
-	FUNCTIONS["md5"] = fnMD5Sum
-	FUNCTIONS["nonempty"] = fnNonEmpty
-	FUNCTIONS["on_path"] = fnOnPath
-	FUNCTIONS["set"] = fnNonEmpty
-	FUNCTIONS["sha1sum"] = fnSha1Sum
-	FUNCTIONS["sha1"] = fnSha1Sum
-	FUNCTIONS["success"] = fnSuccess
-	FUNCTIONS["unset"] = fnEmpty
-	FUNCTIONS["upper"] = fnUpper
 }
